refactor(panpress)!: new default WebApp library
move most generic js from starter to a new 'webapp' library
