--- docs/app/js/app-shell.js
@@ -1,4 +1,0 @@
-import { render } from "lit-html";
-import appShellTemplate from "./components/app-shell.js";
-
-render(appShellTemplate({ lang: "en" }), document.getElementById("app-shell"));
