+++ packages/panpress/starter/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@modern-helpers/lazy-router": "^0.1.0-alpha.3",
+    "@panpress/webapp": "^0.0.0-alpha.2",
     "lit-html": "^1.2.1",
     "prismjs": "^1.20.0"
   }
