+++ packages/panpress/starter/snowpack.config.json
@@ -9,12 +9,17 @@
     "./minify-html-literals.snowpack.js",
     ["@snowpack/plugin-build-script", {"cmd": "postcss", "input": [".css"], "output": [".css"]}]
   ],
-  "installOptions": {},
+  "installOptions": {
+    "treeshake": false
+  },
   "devOptions": {
     "open": "none",
     "out": "dist"
   },
   "buildOptions": {
     "clean": true
+  },
+  "alias": {
+    "~app": "./app"
   }
 }
