+++ packages/panpress/starter/tsconfig.json
@@ -10,7 +10,11 @@
     "strict": false,
     "noImplicitThis": true,
     "alwaysStrict": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "baseUrl": ".",
+    "paths": {
+      "~app/*": ["app/*"]
+    }
   },
   "exclude": [
     "app/views",
