+++ packages/panpress/webapp/components/home.js
@@ -1,7 +1,7 @@
 /**
  * @typedef {{title: string, details: string}} Feature
- * @typedef {import('../../node_modules/lit-html/lib/template-result').TemplateResult} TemplateResult
- * @typedef {{ heroText: string; heroSubText: string; tagline: string; actionText: string; features?: Feature[]; license: string; copyright: string; content?: TemplateResult }} HomeData
+ * @typedef {import('lit-html').TemplateResult} TemplateResult
+ * @typedef {{ heroText: string; heroSubText: string; tagline: string; actionText: string; features?: Feature[]; license: string; copyright: string; logo: TemplateResult; content?: TemplateResult }} HomeData
  */
 
 import { html } from "lit-html";
@@ -145,53 +145,7 @@ export default (data) => html`
     }
   </style>
   <header class="hero">
-    <svg
-      version="1.1"
-      id="Layer_1"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlns:xlink="http://www.w3.org/1999/xlink"
-      x="0px"
-      y="0px"
-      viewBox="0 0 960 960"
-      style="enable-background:new 0 0 960 960;"
-      xml:space="preserve"
-      width="200px"
-    >
-      <style type="text/css">
-        .st0 {
-          fill: #4d4d4d;
-        }
-        .st1 {
-          fill: #ffffff;
-        }
-        .st2 {
-          fill: var(--secondary-color);
-        }
-        .st3 {
-          fill: #333333;
-        }
-      </style>
-      <g>
-        <path
-          class="st0"
-          d="M673.4,766.9H279.3V142.4h394.2c18.8,0,34,15.2,34,34v556.5C707.5,751.7,692.2,766.9,673.4,766.9z"
-        />
-        <path
-          class="st1"
-          d="M661.5,755.6h-379V153.7h379c17.6,0,31.9,14.3,31.9,31.9v538C693.5,741.3,679.2,755.6,661.5,755.6z"
-        />
-        <path
-          class="st0"
-          d="M646.7,766.9H269.6c-9.4,0-17-7.6-17-17V159.4c0-9.4,7.6-17,17-17h377.1c18.8,0,34,15.2,34,34v556.5   C680.7,751.7,665.5,766.9,646.7,766.9z"
-        />
-        <rect x="252.5" y="380.4" class="st2" width="428.2" height="217.4" />
-        <rect x="575.7" y="142.4" class="st3" width="39.1" height="624.6" />
-        <polygon
-          class="st2"
-          points="289.2,766.9 289.2,817.6 314.8,802.4 340.5,817.6 340.5,766.9  "
-        />
-      </g>
-    </svg>
+    ${data.logo}
 
     <h1 id="main-title">
       ${data.heroText}
