+++ packages/panpress/webapp/content/lang.js
@@ -1,0 +1,12 @@
+import { baseUrl, navigate } from "@modern-helpers/lazy-router";
+import { langBase } from "../states/lang.js";
+
+function getGenericPath() {
+  return (
+    window.location.pathname.replace(baseUrl, "").replace(langBase, "/") || "/"
+  );
+}
+
+export async function reload() {
+  return navigate(getGenericPath());
+}
