+++ packages/panpress/webapp/index.js
@@ -1,0 +1,1 @@
+// no default export
