+++ packages/panpress/webapp/main.d.ts
@@ -1,0 +1,1 @@
+// no default export
